test: refactor auth util
