+++ api/test/integration/utils.js
@@ -1,0 +1,17 @@
+export const authorizeUser = async ({
+  credentials,
+  app,
+  setToken,
+}) => {
+  await app.ready();
+
+  const response = await app.inject({
+    method: 'POST',
+    url: '/api/v1/user/signin',
+    payload: credentials,
+  });
+
+  const data = JSON.parse(response.payload);
+
+  setToken(data.accessToken);
+};
